refactor(statistics): create new route to statistics
