+++ src/app/http/statistics/statistics.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+
+import { PatientsModule } from '../patients/patients.module';
+import { StatisticsController } from './statistics.controller';
+import { StatisticsService } from './statistics.service';
+
+@Module({
+  imports: [PatientsModule],
+  controllers: [StatisticsController],
+  providers: [StatisticsService],
+})
+export class StatisticsModule {}
